+++ src/main/java/tech/vitalis/caringu/dtos/TreinoExercicio/TreinoExercicioResponseGetDto.java
@@ -5,13 +5,14 @@ import tech.vitalis.caringu.dtos.Treino.TreinoResponseGetDTO;
 import tech.vitalis.caringu.enums.TreinoExercicio.GrauDificuldadeEnum;
 import tech.vitalis.caringu.enums.TreinoExercicio.OrigemTreinoExercicioEnum;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record TreinoExercicioResponseGetDto(
         Integer id,
         TreinoResponseGetDTO treinos,
         ExercicioResponseGetDTO exercicio,
-        Double carga,
+        BigDecimal carga,
         Integer repeticoes,
         Integer series,
         Integer descanso,
